pyproject.toml
[tool.poetry]
name = "youtube-likes"
version = "0.1.0"
description = ""
authors = ["Hugh Perkins <hughperkins@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
ruamel-yaml = "^0.18.5"
aiohttp = "^3.9.1"
google-api-python-client = "^2.111.0"
google-auth = "^2.25.2"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
chili = "<2.0.0"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
flake8 = "^6.1.0"
black = "^23.12.1"
isort = "^5.13.2"
types-requests = "^2.31.0.10"
flake8-pyproject = "^1.2.3"
google-api-python-client-stubs = "^1.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "E501"]
